feat(IDCIKeycloakSecurityBundle) Working on fix bundle error.
 * Fix circular call on security services from request service.
 * Delete request service class and moved code to security service.
